attempting to move dedicated functions out
